backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
qrcode==7.4.2
python-multipart==0.0.9

sqlmodel==0.0.21
passlib[bcrypt]==1.7.4
python-jose==3.3.0

python-docx==1.1.0

reportlab==4.2.2

pydantic-settings==2.4.0
httpx==0.27.2
